+++ catalog/app/components/Preview/renderers/Perspective/SchemaEditor.tsx
